Translator flask app/requirements.txt
# flask==2.2.2
# azure-core==1.25.0
# azure-ai-textanalytics==5.1.0

azure-ai-textanalytics==5.1.0
azure-common==1.1.28
azure-core==1.25.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
idna==3.3
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msrest==0.7.1
oauthlib==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2